requirements.txt
# Dependencias del host Windows (mt5_executor, precision_sniper, report)
# El container Docker usa su propio requirements.txt en docker/requirements.txt
#
# Instalar con:  pip install -r requirements.txt

MetaTrader5>=5.0.45
numpy>=1.26
pandas>=2.2
flask>=3.1
python-dotenv>=1.0
